Refactored code to more standard way Some Readme updates
