+++ lib/async_button.dart
@@ -1,8 +1,7 @@
 library async_button;
 
-// The following files will be removed in the future versions
-// as the associated classes, functions, etc were deprecated after v1.0.1-beta.2
-export 'src/helpers/constants.dart';
+// The following (and it's associated) files will be removed in the future versions
+// as the exported classes, functions, etc were deprecated after v1.0.1-beta.2
 export 'src/helpers/async_button_state_style.dart';
 export 'src/core/async_button_state_controller_abstract.dart';
 export 'src/async_elevated_button.dart';
@@ -11,13 +10,13 @@ export 'src/async_text_button.dart';
 // ----
 
 // Complimentary
-export 'source/helpers/constants.dart';
-export 'source/helpers/async_btn_state_style.dart';
+export 'src/helpers/constants.dart';
+export 'src/helpers/async_btn_state_style.dart';
 
 // Convenience
-export 'source/core/async_btn_states_controller.dart';
+export 'src/core/async_btn_states_controller.dart';
 
 // Necessary
-export 'source/async_elevated_btn.dart';
-export 'source/async_outlined_btn.dart';
-export 'source/async_text_btn.dart';
+export 'src/async_elevated_btn.dart';
+export 'src/async_outlined_btn.dart';
+export 'src/async_text_btn.dart';
